Refactor AppNavigator to have Activity impl
