+++ app/src/main/java/me/banes/chris/tivi/details/ShowDetailsModule.kt
@@ -14,22 +14,17 @@
  * limitations under the License.
  */
 
-package me.banes.chris.tivi.home.trending
+package me.banes.chris.tivi.details
 
-import android.arch.lifecycle.ViewModel
-import dagger.Binds
 import dagger.Module
-import dagger.android.ContributesAndroidInjector
-import dagger.multibindings.IntoMap
-import me.banes.chris.tivi.inject.ViewModelKey
+import dagger.Provides
+import me.banes.chris.tivi.AppNavigator
+import me.banes.chris.tivi.TiviAppActivityNavigator
 
 @Module
-internal abstract class TrendingBuilder {
-    @ContributesAndroidInjector
-    internal abstract fun trendingShowsFragment(): TrendingShowsFragment
-
-    @Binds
-    @IntoMap
-    @ViewModelKey(TrendingShowsViewModel::class)
-    abstract fun bindTrendingShowsViewModel(viewModel: TrendingShowsViewModel): ViewModel
+class ShowDetailsModule {
+    @Provides
+    fun provideAppNavigator(activity: ShowDetailsActivity): AppNavigator {
+        return TiviAppActivityNavigator(activity)
+    }
 }
