+++ app/src/main/java/me/banes/chris/tivi/home/discover/DiscoverViewModel.kt
@@ -92,6 +92,6 @@ internal class DiscoverViewModel @Inject constructor(
     }
 
     fun onItemPostedClicked(navigator: HomeNavigator, show: TiviShow) {
-        navigator.showShowDetails(show)
+        navigator.showShowDetails(show, null) // TODO
     }
 }
