+++ app/src/main/java/me/banes/chris/tivi/home/library/LibraryViewModel.kt
@@ -78,7 +78,7 @@ class LibraryViewModel @Inject constructor(
     }
 
     fun onItemPostedClicked(navigator: HomeNavigator, show: TiviShow) {
-        navigator.showShowDetails(show)
+        navigator.showShowDetails(show, null) // TODO
     }
 
 }
