+++ app/src/main/res/layout/fragment_show_details.xml
@@ -73,6 +73,6 @@
         android:layout_height="120dp"
         android:layout_marginStart="16dp"
         android:elevation="2dp"
-        app:layout_behavior="me.banes.chris.tivi.ui.PosterBehavior" />
+        app:layout_behavior="me.banes.chris.tivi.details.DetailPosterBehavior" />
 
 </android.support.design.widget.CoordinatorLayout>
